+++ src/main/resources/application-dev.yml
@@ -8,7 +8,7 @@ spring:
   jpa:
     open-in-view: true
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
       naming:
         physical-strategy: com.example.howmuch.config.app.UpperCaseNamingStrategy
     show-sql: true
